Fix issue where restaurant stars wont update with new review

+++ src/main/java/ca/ece/ubc/cpen221/mp5/database/YelpRestaurant.java
@@ -129,6 +129,8 @@ public class YelpRestaurant implements Business{
 	return restaurantJson.toString();
     }
 
+    
+    // NOTE DOES NOT UPDATE RATING 
     public void addReview(String review) {
 	reviews.add(review);
 	reviewCount ++;
@@ -199,7 +201,7 @@ public class YelpRestaurant implements Business{
 	return categories.remove(category);
     }
 
-    public void setStars(int stars) {
+    public void setStars(double stars) {
 	this.stars = stars;
     }
